Remove PlaybackService from the library
